+++ Plutarch/Api/V1/Time.hs
@@ -2,7 +2,7 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Plutarch.Api.V1.Time (
-  PPOSIXTime,
+  PPOSIXTime (PPOSIXTime),
   PPOSIXTimeRange,
 ) where
 
@@ -11,6 +11,7 @@ import qualified Plutus.V1.Ledger.Api as Plutus
 import Plutarch.Api.V1.Interval (PInterval)
 import Plutarch.Lift (
   DerivePConstantViaNewtype (DerivePConstantViaNewtype),
+  PConstantDecl,
   PLifted,
   PUnsafeLiftDecl,
  )
@@ -20,10 +21,12 @@ newtype PPOSIXTime (s :: S)
   = PPOSIXTime (Term s PInteger)
   deriving (PlutusType, PIsData, PEq, POrd, PIntegral) via (DerivePNewtype PPOSIXTime PInteger)
 
+deriving via (Term s (DerivePNewtype PPOSIXTime PInteger)) instance Num (Term s PPOSIXTime)
+
 instance PUnsafeLiftDecl PPOSIXTime where type PLifted PPOSIXTime = Plutus.POSIXTime
 deriving via
   (DerivePConstantViaNewtype Plutus.POSIXTime PPOSIXTime PInteger)
   instance
-    (PConstant Plutus.POSIXTime)
+    PConstantDecl Plutus.POSIXTime
 
 type PPOSIXTimeRange = PInterval PPOSIXTime