+++ Plutarch/Lift.hs
@@ -3,6 +3,14 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 
+{- |
+Module: Plutarch.Lift
+Description: Conversion to and from Plutarch terms and Haskell types
+
+This module defines functions, associated type families, and newtypes for use with
+[@DerivingVia@](https://ryanglscott.github.io/papers/deriving-via.pdf) to allow
+Plutarch to convert to and from PTypes and Haskell types.
+-}
 module Plutarch.Lift (
   -- * Converstion between Plutarch terms and Haskell types
   pconstant,
@@ -11,8 +19,9 @@ module Plutarch.Lift (
   LiftError,
 
   -- * Define your own conversion
-  PConstant (..),
+  PConstantDecl (..),
   PLift,
+  PConstant,
   DerivePConstantDirect (..),
   DerivePConstantViaNewtype (..),
   DerivePConstantViaBuiltin (..),
@@ -22,8 +31,8 @@ module Plutarch.Lift (
 ) where
 
 import Control.Lens ((^?))
-import Data.Coerce
-import Data.Kind (Type)
+import Data.Coerce (Coercible, coerce)
+import Data.Kind (Constraint, Type)
 import GHC.Stack (HasCallStack)
 import Plutarch.Evaluate (EvalError, evalScript)
 import Plutarch.Internal (ClosedTerm, PType, Term, compile, punsafeConstantInternal)
@@ -40,8 +49,8 @@ Laws:
  - It must be that @PConstantRepr (PLifted p)@ when encoded as a constant
    in UPLC (via the 'UntypedPlutusCore.Constant' constructor) is a valid @p@.
 -}
-class (PConstant (PLifted p), PConstanted (PLifted p) ~ p) => PUnsafeLiftDecl (p :: PType) where
-  type PLifted p :: Type
+class (PConstantDecl (PLifted p), PConstanted (PLifted p) ~ p) => PUnsafeLiftDecl (p :: PType) where
+  type PLifted p = (r :: Type) | r -> p
 
 {- | Class of Haskell types `h` that can be represented as a Plutus core builtin
 and converted to a Plutarch type.
@@ -58,7 +67,12 @@ Laws:
 
 These laws must be upheld for the sake of soundness of the type system.
 -}
-class (PUnsafeLiftDecl (PConstanted h), PLC.DefaultUni `PLC.Includes` PConstantRepr h) => PConstant (h :: Type) where
+class
+  ( PUnsafeLiftDecl (PConstanted h)
+  , PLC.DefaultUni `PLC.Includes` PConstantRepr h
+  ) =>
+  PConstantDecl (h :: Type)
+  where
   type PConstantRepr h :: Type
   type PConstanted h :: PType
   pconstantToRepr :: h -> PConstantRepr h
@@ -70,6 +84,7 @@ The Haskell type is determined by `PLifted p`.
 
 This typeclass is closely tied with 'PConstant'.
 -}
+type PLift :: PType -> Constraint
 type PLift = PUnsafeLiftDecl
 
 {- | Create a Plutarch-level constant, from a Haskell value.
@@ -113,22 +128,72 @@ plift prog = case plift' prog of
             <> maybe "" (\x -> "cause: " <> show x) causeMaybe
   Left (LiftError_EvalError e) -> error $ "plift failed: erring term: " <> show e
 
--- TODO: Add haddock
+{- | Newtype wrapper for deriving @PConstant@ when the wrapped type is directly
+represented by a builtin UPLC type that is /not/ @Data@.
+
+  Ex: @PInteger@ is directly represented as a builtin integer.
+-}
 newtype DerivePConstantDirect (h :: Type) (p :: PType) = DerivePConstantDirect h
 
 instance
   (PLift p, PLC.DefaultUni `PLC.Includes` h) =>
-  PConstant (DerivePConstantDirect h p)
+  PConstantDecl (DerivePConstantDirect h p)
   where
   type PConstantRepr (DerivePConstantDirect h p) = h
   type PConstanted (DerivePConstantDirect h p) = p
   pconstantToRepr = coerce
   pconstantFromRepr = Just . coerce
 
--- TODO: Add haddock
-newtype DerivePConstantViaNewtype (h :: Type) (p :: PType) (p' :: PType) = DerivePConstantViaNewtype h
+{- | Newtype wrapper for deriving @PConstant@ when the wrapped type is represented
+indirectly by a builtin UPLC type that is /not/ @Data@.
 
-instance (PLift p, PLift p', Coercible h (PLifted p')) => PConstant (DerivePConstantViaNewtype h p p') where
+  Ex: @PPubKeyHash@ is a newtype to a @PByteString@ and @PByteString@ is directly
+  represented as a builtin bytestring.
+
+Polymorphic types can be derived as follows:
+
+>newtype Foo a = Foo a
+>
+>newtype PFoo a s = PFoo (Term s a)
+>
+>instance forall a. PLift a => PUnsafeLiftDecl (PFoo a) where
+>  type PLifted (PFoo a) = Foo (PLifted a)
+>
+>deriving via
+>  ( DerivePConstantViaNewtype
+>      (Foo a)
+>      (PFoo (PConstanted a))
+>      (PConstanted a)
+>  )
+>  instance
+>    PConstant a =>
+>    PConstantDecl (Foo a)
+-}
+newtype
+  DerivePConstantViaNewtype
+    (h :: Type)
+    (p :: PType) -- PType to associate with the newtype
+    (p' :: PType) -- Underlying UPLC representation type
+  = -- | The Haskell newtype we are deriving a @PConstant@ instance for
+    DerivePConstantViaNewtype h
+
+{- | Type synonym to simplify deriving of @PConstant@ via @DerivePConstantViaNewtype@.
+
+A newtype @Foo a@ is considered "Constantable" if:
+
+- The wrapped type @a@ has a @PConstant@ instance.
+- The lifted type of @a@ has a @PUnsafeLiftDecl@ instance.
+- There is type equality between @a@ and @PLifted (PConstanted a)@.
+
+These constraints are sufficient to derive a @PConstant@ instance for the newtype.
+
+For deriving @PConstant@ for a wrapped type represented in UPLC as @Data@, see
+@DerivePConstantViaData@.
+-}
+type PConstant :: Type -> Constraint
+type PConstant a = (a ~ PLifted (PConstanted a), PConstantDecl a)
+
+instance (PLift p, PLift p', Coercible h (PLifted p')) => PConstantDecl (DerivePConstantViaNewtype h p p') where
   type PConstantRepr (DerivePConstantViaNewtype h p p') = PConstantRepr (PLifted p')
   type PConstanted (DerivePConstantViaNewtype h p p') = p
   pconstantToRepr x = pconstantToRepr @(PLifted p') $ coerce x
@@ -154,7 +219,15 @@ instance FromBuiltin' BuiltinData Data where
 
 newtype DerivePConstantViaBuiltin (h :: Type) (p :: PType) (p' :: PType) = DerivePConstantViaBuiltin h
 
-instance (PLift p, PLift p', Coercible h h', ToBuiltin' (PLifted p') h', FromBuiltin' h' (PLifted p')) => PConstant (DerivePConstantViaBuiltin h p p') where
+instance
+  ( PLift p
+  , PLift p'
+  , Coercible h h'
+  , ToBuiltin' (PLifted p') h'
+  , FromBuiltin' h' (PLifted p')
+  ) =>
+  PConstantDecl (DerivePConstantViaBuiltin h p p')
+  where
   type PConstantRepr (DerivePConstantViaBuiltin h p p') = PConstantRepr (PLifted p')
   type PConstanted (DerivePConstantViaBuiltin h p p') = p
   pconstantToRepr x = pconstantToRepr @(PLifted p') $ fromBuiltin' (coerce x :: h')