+++ Plutarch/Prelude.hs
@@ -118,8 +118,8 @@ module Plutarch.Prelude (
   pdcons,
   pdnil,
   pfield,
+  getField,
   pletFields,
-  hrecField,
 
   -- * Tracing
   ptrace,
@@ -140,6 +140,8 @@ module Plutarch.Prelude (
   plift,
   PConstant,
   PLift,
+  PConstantData,
+  PLiftData,
 
   -- * Typeclass derivers.
   DerivePNewtype (DerivePNewtype),
@@ -153,6 +155,7 @@ module Plutarch.Prelude (
 import Prelude ()
 
 import Data.Kind (Type)
+import GHC.Records (getField)
 import Plutarch
 import Plutarch.Bool
 import Plutarch.Builtin